cleaning the code and some config files
