+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_11_k_ref_values_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 3.829964341609693,
+    "1": 3.7260099144414247,
+    "2": 3.8241834318474948,
+    "3": 3.6484909600970656,
+    "4": 3.901575202236645,
+    "5": 3.6899083236811556,
+    "6": 3.746635108069917,
+    "7": 3.8383602895029516,
+    "8": 3.9363249981489004,
+    "9": 3.5723152168835117,
+    "10": 3.9928001577094943,
+    "11": 3.528477884503721,
+    "12": 4.228181219651002,
+    "13": 6.896131468229814,
+    "14": 5.979840558059788,
+    "15": 3.4938390839652707,
+    "16": 5.6407509254441965,
+    "17": 3.300980695108981,
+    "18": 6.595724117172372,
+    "19": 8.159631171064362,
+    "20": 3.9430343429004457,
+    "21": 3.704945754782311,
+    "22": 3.8195878653118207,
+    "23": 3.4214306441300684
+}
