+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_11_ref_log_mean_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 0.5863270682559714,
+    "1": 0.5747339900355345,
+    "2": 0.5954957730744972,
+    "3": 0.6066260470698204,
+    "4": 0.559788834072985,
+    "5": 0.5617170763765513,
+    "6": 0.5605280752353207,
+    "7": 0.5836255810980872,
+    "8": 0.5744388989204361,
+    "9": 0.567537996318177,
+    "10": 0.5662247711827358,
+    "11": 0.574065558915337,
+    "12": 0.6482084097141279,
+    "13": 0.5935320137141189,
+    "14": 0.5921700380153118,
+    "15": 0.5959472521856262,
+    "16": 0.6043274272867106,
+    "17": 0.5978632609598188,
+    "18": 0.5986499443200569,
+    "19": 0.5975218253501201,
+    "20": 0.6004076217418405,
+    "21": 0.5881029823964978,
+    "22": 0.5991752508608613,
+    "23": 0.5788986097700908
+}
