+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_11_ref_log_std_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 0.039048284145081956,
+    "1": 0.038466682480403705,
+    "2": 0.042426582128760555,
+    "3": 0.044282756478084766,
+    "4": 0.03789374277450213,
+    "5": 0.03708977112174703,
+    "6": 0.0373502594703324,
+    "7": 0.04068163991446997,
+    "8": 0.03861414062867147,
+    "9": 0.03750727636076541,
+    "10": 0.03793662578186129,
+    "11": 0.03799720643430259,
+    "12": 0.14188317203806913,
+    "13": 0.0491746661637587,
+    "14": 0.046034936132517315,
+    "15": 0.04302160650199867,
+    "16": 0.04968050141483007,
+    "17": 0.04235110281204008,
+    "18": 0.048375498495275844,
+    "19": 0.049507713534904546,
+    "20": 0.04208466977318288,
+    "21": 0.03872332802014452,
+    "22": 0.04114908520822918,
+    "23": 0.038480977755699024
+}
