+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_12_k_ref_values_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 3.849795013849453,
+    "1": 4.255796310136959,
+    "2": 3.557146173514604,
+    "3": 3.1120621748050303,
+    "4": 3.658991398922919,
+    "5": 3.704100630732048,
+    "6": 4.009864481222621,
+    "7": 3.524649935913531,
+    "8": 3.1865783569151405,
+    "9": 3.6110022371777437,
+    "10": 3.7052693402498615,
+    "11": 3.814422512226967,
+    "12": 3.5403832487989066,
+    "13": 3.1768323066979414,
+    "14": 3.2127529699668367,
+    "15": 2.96451969844121,
+    "16": 4.050907342306148,
+    "17": 3.3970855424858093,
+    "18": 3.14051136490855,
+    "19": 3.631191109202381,
+    "20": 3.1603440092920048,
+    "21": 4.171886725593125,
+    "22": 4.194087558699685,
+    "23": 4.556853674149407
+}
