+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_12_ref_log_mean_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 0.5713513249937787,
+    "1": 0.567984208145071,
+    "2": 0.5977074039008049,
+    "3": 0.570767373360953,
+    "4": 0.5559490310451752,
+    "5": 0.5605373229710555,
+    "6": 0.563731896239067,
+    "7": 0.5646156984924069,
+    "8": 0.5550939691253318,
+    "9": 0.5787937216975721,
+    "10": 0.5529324042225692,
+    "11": 0.5595844143884696,
+    "12": 0.5931353902058459,
+    "13": 0.5946646016048273,
+    "14": 0.5898895695197092,
+    "15": 0.5950852649572446,
+    "16": 0.5896709360542118,
+    "17": 0.5915678991773604,
+    "18": 0.58895126470135,
+    "19": 0.5950518954456222,
+    "20": 0.5994004127477096,
+    "21": 0.5624225767993091,
+    "22": 0.5781501464895522,
+    "23": 0.5587261762602304
+}
