+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_12_ref_log_std_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 0.039383837247100116,
+    "1": 0.0370559564071702,
+    "2": 0.04239536780886001,
+    "3": 0.037589526479226094,
+    "4": 0.03659148531477169,
+    "5": 0.035754882047784875,
+    "6": 0.03645055134724188,
+    "7": 0.0386351610765656,
+    "8": 0.03573113556987117,
+    "9": 0.038413828044258344,
+    "10": 0.0345483183342533,
+    "11": 0.03632377805783317,
+    "12": 0.04059712776930658,
+    "13": 0.04171805637527494,
+    "14": 0.04139830131023769,
+    "15": 0.041378560168600656,
+    "16": 0.04227318731628883,
+    "17": 0.03892083682070472,
+    "18": 0.04013164666008581,
+    "19": 0.041453032483207274,
+    "20": 0.03966529621934177,
+    "21": 0.036237784186349827,
+    "22": 0.06354086623734673,
+    "23": 0.036548414423939087
+}
