+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_13_k_ref_values_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 3.900679495850332,
+    "1": 3.200975559140311,
+    "2": 4.361277838883146,
+    "3": 4.062852150609478,
+    "4": 3.75984587103894,
+    "5": 3.6653823447383376,
+    "6": 4.188397455315481,
+    "7": 3.6805898708207776,
+    "8": 3.6850534384385853,
+    "9": 3.7972818023593664,
+    "10": 3.5787903884706194,
+    "11": 4.087885584675547,
+    "12": 3.510276400715728,
+    "13": 4.73702096550692,
+    "14": 3.779773894218344,
+    "15": 3.563366868637202,
+    "16": 3.989123214749204,
+    "17": 3.8260076986593012,
+    "18": 3.9331916723763554,
+    "19": 3.5393370031069677,
+    "20": 3.594570039520564,
+    "21": 3.554681560606197,
+    "22": 3.415408657587062,
+    "23": 3.688094555330375
+}
