+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_13_ref_log_mean_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 0.6194456623256833,
+    "1": 0.5937474383811342,
+    "2": 0.5973403688774467,
+    "3": 0.584232418263017,
+    "4": 0.5730654218513959,
+    "5": 0.5580363727640998,
+    "6": 0.566888875198597,
+    "7": 0.5581038928072259,
+    "8": 0.566857174525873,
+    "9": 0.5622526312833439,
+    "10": 0.5712216111445392,
+    "11": 0.5758386780125822,
+    "12": 0.5992572598361109,
+    "13": 0.5978135967126665,
+    "14": 0.5954534460932793,
+    "15": 0.5981637891430953,
+    "16": 0.6048134701127879,
+    "17": 0.5965119406956066,
+    "18": 0.5930343323302318,
+    "19": 0.6043670953408035,
+    "20": 0.5971162836158682,
+    "21": 0.5628175935185491,
+    "22": 0.5773909241351641,
+    "23": 0.5640616252857331
+}
