+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_13_ref_log_std_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 0.051836271544717843,
+    "1": 0.039477516511881854,
+    "2": 0.10066676477660347,
+    "3": 0.039144360137740676,
+    "4": 0.03797295676736169,
+    "5": 0.036854250929583894,
+    "6": 0.03845371749281073,
+    "7": 0.03656540626526353,
+    "8": 0.03722473595589377,
+    "9": 0.037271852841641355,
+    "10": 0.03789143996059949,
+    "11": 0.03808893532616364,
+    "12": 0.040267249812766026,
+    "13": 0.04706818587590845,
+    "14": 0.03946814443185916,
+    "15": 0.04147701650933867,
+    "16": 0.04256714027296742,
+    "17": 0.04181186534619488,
+    "18": 0.0403017662986866,
+    "19": 0.04324461213945535,
+    "20": 0.04089416723547774,
+    "21": 0.03679523998142298,
+    "22": 0.038509725220709064,
+    "23": 0.03797611470296361
+}
