+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_23_k_ref_values_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 3.3496578007138074,
+    "1": 3.947155240549477,
+    "2": 3.671511992611104,
+    "3": 3.453529271807341,
+    "4": 3.730515529242998,
+    "5": 3.7636242827320596,
+    "6": 3.5589704983542823,
+    "7": 3.7179246006692552,
+    "8": 3.7402291977349003,
+    "9": 3.60633029323059,
+    "10": 3.4717746730450147,
+    "11": 3.9374437405114557,
+    "12": 3.5911746204952366,
+    "13": 3.6222310406478453,
+    "14": 3.6723101839478023,
+    "15": 3.4441001361897454,
+    "16": 4.019640585866586,
+    "17": 3.7105927441249276,
+    "18": 3.640636387919619,
+    "19": 3.8852774499665865,
+    "20": 3.845517646476147,
+    "21": 3.832140672818369,
+    "22": 3.651766130366839,
+    "23": 3.5423371296995207
+}
