+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_23_ref_log_mean_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 0.5781494372817108,
+    "1": 0.5682401524183903,
+    "2": 0.5661317767476333,
+    "3": 0.5883592329584293,
+    "4": 0.5634263539299192,
+    "5": 0.5886873779551022,
+    "6": 0.5819363732147549,
+    "7": 0.59578243394305,
+    "8": 0.5656352854331792,
+    "9": 0.5632163281328884,
+    "10": 0.5610397674045451,
+    "11": 0.571783208367988,
+    "12": 0.599718469559197,
+    "13": 0.6038609732062177,
+    "14": 0.596357611757548,
+    "15": 0.5961644957650608,
+    "16": 0.6010846446659442,
+    "17": 0.5985812662221962,
+    "18": 0.5992886048121578,
+    "19": 0.5992602312791105,
+    "20": 0.5996133920835652,
+    "21": 0.5733108596721112,
+    "22": 0.5604074060788465,
+    "23": 0.5710027430257861
+}
