+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/expected_values/station_23_ref_log_std_snr.json
@@ -1,0 +1,26 @@
+{
+    "0": 0.03819835044437628,
+    "1": 0.03719133277057184,
+    "2": 0.03667775600086205,
+    "3": 0.039302608355263925,
+    "4": 0.03736373706002523,
+    "5": 0.0393594029019399,
+    "6": 0.03691624275389805,
+    "7": 0.041282115503721815,
+    "8": 0.03679655891686211,
+    "9": 0.03745650611723401,
+    "10": 0.03737772921364614,
+    "11": 0.03818390159664768,
+    "12": 0.04139535291126791,
+    "13": 0.042057195578260535,
+    "14": 0.04048560218219939,
+    "15": 0.041446398313918345,
+    "16": 0.04165690796683088,
+    "17": 0.04072061047731856,
+    "18": 0.040752989415819854,
+    "19": 0.04082098167868109,
+    "20": 0.039869956050217335,
+    "21": 0.03823961627886271,
+    "22": 0.036876286579642065,
+    "23": 0.03715641376820358
+}
