+++ rnog_analysis_tools/data_monitoring/science_verification_analysis/science_verification_analysis.py
@@ -1,3 +1,7 @@
+'''
+This module can be used to test if the stations are working as expected.
+'''
+
 from NuRadioReco.modules.io.RNO_G.readRNOGDataMattak import readRNOGData
 import rnog_data.runtable as rt
 from NuRadioReco.detector import detector
@@ -33,150 +37,42 @@ from matplotlib import colors, cm
 from NuRadioReco.modules.RNO_G.channelBlockOffsetFitter import fit_block_offsets
 import builtins
 import csv
+from config_station import get_station_config
+from config_plotting import set_plot_style
+from read_rnog_data_nuradio import convert_events_information, read_rnog_data
+from spectral_analysis_sva import find_amplitude_ratio_in_band, find_amplitude_ratio_in_band_specific_bkg, excess_info_from_ratio, excess_info_from_ratio_specific_bkg, validate_excess_in_bands
+from config_spectral_analysis import SPECTRAL_BANDS, ALPHA_SPEC, CI_THRESHOLDS_SPEC
+import copy
 
-'''
-This module can be used to test if the stations are working as expected.
-'''
+# Logging
+logger = logging.getLogger(__name__)
 
-# Channel mapping for the first seven RNO-G stations:
-SURFACE_CHANNELS_LIST = [12, 13, 14, 15, 16, 17, 18 , 19 , 20]
-DEEP_CHANNELS_LIST = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 21, 22, 23]
-UPWARD_CHANNELS_LIST = [13, 16, 19]
-DOWNWARD_CHANNELS_LIST = [12, 14, 15, 17, 18, 20]
-VPOL_LIST = [0, 1, 2, 3, 5, 6, 7, 9, 10, 22, 23]
-HPOL_LIST = [4, 8, 11, 21]
-PHASED_ARRAY_LIST = [0, 1, 2, 3]
-ALL_CHANNELS_LIST = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18 , 19 , 20, 21, 22, 23]
+def setup_logging(station_id, run_label):
+    os.makedirs("logs", exist_ok=True)
 
-# Channel mapping for Station 14:
-STATION_14_SURFACE_CHANNELS_LIST = [12, 13, 14, 15, 16, 17, 18 , 19]
-STATION_14_DEEP_CHANNELS_LIST = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 20, 21, 22, 23]
-STATION_14_UPWARD_CHANNELS_LIST = [13, 15, 16, 18]
-STATION_14_DOWNWARD_CHANNELS_LIST = [12, 14, 17, 19]
-STATION_14_VPOL_LIST = [0, 1, 2, 3, 5, 6, 7, 9, 10, 20, 22, 23]
-STATION_14_HPOL_LIST = [4, 8, 11, 21]
-STATION_14_PHASED_ARRAY_LIST = [0, 1, 2, 3]
-STATION_14_ALL_CHANNELS_LIST = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18 , 19 , 20, 21, 22, 23]
+    log_file = os.path.join("logs", f"science_verification_analysis_station{station_id}_{run_label}.log")
 
-# Script directory
+    logging.basicConfig(
+        level=logging.INFO,
+        format="%(asctime)s [%(levelname)s] %(message)s",
+        handlers=[
+            logging.FileHandler(log_file, mode="w"),
+            logging.StreamHandler()
+        ],
+        force=True
+    )
+
+    logger.info(f"Logging to {log_file}")
+
+# Script directory for json files
 SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
 
-# Matplotlib settings
-
-COLORS = [
-    "#4477AA",  # strong blue
-    "#EE6677",  # coral / red
-    "#228833",  # green
-    "#CCBB44",  # yellow
-    "#66CCEE",  # light blue
-    "#AA3377",  # purple
-
-    "#882255",  # wine
-    "#44AA99",  # teal
-    "#084C1F",  # deep green
-    "#332288",  # indigo
-    "#AA4499",  # magenta
-    "#771122",  # burgundy
-
-    "#7089A1",  # steel blue
-    "#DDCC77",  # sand
-    "#B0D8EC",  # pale blue
-    "#CC6677",  # dusty red
-    "#999933",  # olive
-    "#DCB43C",  # warm yellow
-
-    "#006699",  # dark cyan
-    "#0099CC",  # vivid cyan
-    "#9955AA",  # lavender purple
-    "#55AA55",  # bright green
-    "#CC7711",  # warm orange
-    "#555555",  # neutral gray
-]
-
-mpl.rcParams.update({
-    'font.family': 'sans-serif',
-    'font.sans-serif': ['Helvetica', 'Arial', 'DejaVu Sans'],
-    'font.size': 17,
-
-    'axes.labelsize': 17,
-    'axes.titlesize': 18,
-    'axes.linewidth': 1.2,
-    'axes.grid': False,
-
-    'axes.prop_cycle': cycler('color', COLORS),
-
-    'xtick.labelsize': 17,
-    'ytick.labelsize': 17,
-    'xtick.major.size': 6,
-    'ytick.major.size': 6,
-    'xtick.major.width': 1.2,
-    'ytick.major.width': 1.2,
-    'xtick.minor.size': 3,
-    'ytick.minor.size': 3,
-    'xtick.minor.visible': True,
-    'ytick.minor.visible': True,
-
-    'lines.linewidth': 1.6,
-    'lines.antialiased': True,
-    'lines.markersize': 6,
-
-    'legend.fontsize': 14,
-    'legend.frameon': False,
-    'legend.handlelength': 1,
-    'legend.borderpad': 0.3,
-
-    'figure.dpi': 120,
-    'savefig.dpi': 300,
-    'savefig.bbox': 'tight',
-})
-
-_print = builtins.print
-_log_file = None
-
-def start_print_logging(path):
-    global _log_file
-    _log_file = open(path, "w")
-
-    def logged_print(*args, **kwargs):
-        _print(*args, **kwargs)  # still print to terminal
-        if _log_file is not None:
-            _print(*args, **kwargs, file=_log_file)
-
-    builtins.print = logged_print
-
-def stop_print_logging():
-    global _log_file
-    if _log_file is not None:
-        _log_file.close()
-        _log_file = None
-    builtins.print = _print
-
-def convert_events_information(event_info, convert_to_arrays=True):
-
-    data = defaultdict(list)
-
-    for ele in event_info.values():
-        for k, v in ele.items():
-            data[k].append(v)
-
-    if convert_to_arrays:
-        for k in data:
-            data[k] = np.array(data[k])
-
-    return data
-
+# Choose events based on trigger type
 def choose_trigger_type(event_info, trigger_type: str):
     '''Choose events based on trigger type.'''
     mask = event_info["triggerType"] == trigger_type
 
     return mask
-    
-# Find the correct channel mapping for a given station
-def station_channels(station_id: int):
-    if station_id == 14:
-        return STATION_14_SURFACE_CHANNELS_LIST, STATION_14_DEEP_CHANNELS_LIST, STATION_14_UPWARD_CHANNELS_LIST, STATION_14_DOWNWARD_CHANNELS_LIST, STATION_14_VPOL_LIST, STATION_14_HPOL_LIST, STATION_14_PHASED_ARRAY_LIST, STATION_14_ALL_CHANNELS_LIST
-    else:
-        return SURFACE_CHANNELS_LIST, DEEP_CHANNELS_LIST, UPWARD_CHANNELS_LIST, DOWNWARD_CHANNELS_LIST, VPOL_LIST, HPOL_LIST, PHASED_ARRAY_LIST, ALL_CHANNELS_LIST
 
 # Normalize surface channel spectra to the average of the down channels in the reference frequency band (500-650 MHz)
 def normalize_channels(spec_arr, frequencies, down_channels, up_channels, f_low=500*units.MHz, f_high=650*units.MHz):
@@ -207,240 +103,6 @@ def read_rnog_runtable(station_id: int, start_time: str, stop_time: str):
     testrt = RunTable.get_table( start_time=start_time, stop_time=stop_time, stations=[station_id], run_types = ['physics'])
     return testrt
 
-def read_rnog_data(station_id: int, run_numbers: list, backend: str = "pyroot"):
-    '''Read RNO-G data for a given station and list of run numbers using the specified backend.'''
-    file_list = []
-    valid_run_numbers = []
-    missing_runs = []
-
-    for run_id in run_numbers:
-        path = f"/pnfs/ifh.de/acs/radio/diskonly/data/inbox/station{station_id}/run{run_id}/combined.root"
-        if os.path.isfile(path):
-            file_list.append(path)
-            valid_run_numbers.append(run_id)
-        else:
-            missing_runs.append(run_id)
-
-    if missing_runs:
-        print(f"!!!! Skipping {len(missing_runs)} missing runs: {missing_runs} !!!!")
-    if not file_list:
-        raise FileNotFoundError("No combined.root files found for selected runs.")
-    
-    n_files = len(file_list)
-    n_batches = n_files // 100 + 1
-    print(f"Reading {n_files} files in {n_batches} batches using {backend} backend.")
-
-    event_info = defaultdict(list)
-
-    n_events_total = 0
-    spec_batches = []
-    trace_batches = []
-    times_trace_batches = []
-    snr_batches = []
-    glitch_batches = []
-    run_no_all = []
-    times_all = []
-    freqs = None
-
-    from NuRadioReco.modules.channelSignalReconstructor import channelSignalReconstructor as csr
-    csr = csr()
-    csr.begin(debug=False)
-
-    from NuRadioReco.modules.RNO_G.channelGlitchDetector import channelGlitchDetector as cgd_rnog
-    cgd_rnog = cgd_rnog()
-    cgd_rnog.begin()
-
-    for batch in tqdm(np.array_split(np.array(file_list), n_batches), desc="Reading batches", unit="batch"):
-        tableReader = dataProviderRNOG()
-        tableReader.begin(files=batch.tolist(), 
-                          det=None,
-                          reader_kwargs={"overwrite_sampling_rate":2.4, 
-                                         "convert_to_voltage":True,
-                                         "apply_baseline_correction":"auto",
-                                         "mattak_kwargs":{"backend":backend}})
-        event_info_tmp = tableReader.reader.get_events_information(
-            keys=["triggerType", "triggerTime", "readoutTime", "radiantThrs", "lowTrigThrs", "run", "eventNumber"])
-        
-        event_info_tmp = convert_events_information(event_info_tmp, False)
-        for key, value in event_info_tmp.items():
-            event_info[key] += value
-
-        n_events = tableReader.reader.get_n_events()
-        n_events_total += n_events
-        print(f"Reading {n_events} events in this batch.")
-
-        channel_list = [i for i in range(24)]  
-        spec_arr = np.zeros((len(channel_list), n_events, 1025))
-        trace_arr = np.zeros((len(channel_list), n_events, 2048))
-        times_trace_arr = np.zeros((len(channel_list), n_events, 2048))
-        snr_arr = np.zeros((len(channel_list), n_events))
-        glitch_arr = np.zeros((len(channel_list), n_events))
-
-        run_no = []
-        times = []
-        event_ids = []
-        
-        for idx, event in enumerate(tqdm(tableReader.reader.run(), total=n_events, desc="Events", unit="evt", leave=False)):
-            station = event.get_station()
-            time = station.get_station_time().datetime64
-            times.append(time)
-            run_no.append(event.get_run_number())
-
-            csr.run(evt=event, station=station, det=None, stored_noise=False)
-            cgd_rnog.run(event=event, station=station, det=None)
-            for i_ch, ch in enumerate(channel_list):
-                channel = station.get_channel(ch)
-
-                times_ch = channel.get_times()
-                times_trace_arr[i_ch, idx, :] = times_ch
-
-                snr_dict = channel.get_parameter(chp.SNR)
-                snr_peak = snr_dict["peak_amplitude"]
-                snr_arr[i_ch, idx] = snr_peak
-
-                glitching_values = channel.get_parameter(chp_rnog.glitch_test_statistic)
-                glitch_arr[i_ch, idx] = glitching_values
-                
-                spec = channel.get_frequency_spectrum()
-                spec_arr[i_ch, idx, :] = np.abs(spec)
-
-                trace = channel.get_trace()
-                trace_arr[i_ch, idx, :] = trace
-
-                if freqs is None and idx == 0 and i_ch == 0:
-                    freqs = channel.get_frequencies()
-        
-        spec_batches.append(spec_arr)
-        trace_batches.append(trace_arr)
-        times_trace_batches.append(times_trace_arr)
-        snr_batches.append(snr_arr)
-        glitch_batches.append(glitch_arr)
-        run_no_all.extend(run_no)
-        times_all.extend(times)
-
-    spec_arr = np.concatenate(spec_batches, axis=1)
-    trace_arr = np.concatenate(trace_batches, axis=1)
-    times_trace_arr = np.concatenate(times_trace_batches, axis=1)
-    snr_arr = np.concatenate(snr_batches, axis=1)
-    glitch_arr = np.concatenate(glitch_batches, axis=1)
-
-    run_no = np.array(run_no_all)
-    times = np.array(times_all)     
-
-    for key, value in event_info.items():
-        event_info[key] = np.array(value)
-
-    inf_mask = np.isinf(event_info["triggerTime"])
-    event_info["triggerTime"][inf_mask] = event_info["readoutTime"][inf_mask]
-    print(f"Found {np.sum(inf_mask)} events with inf trigger time (of {len(inf_mask)} events)")
-
-    print(f"n_events read: {spec_arr.shape[1]}, n_events_total: {n_events_total}")
-    print(f"freqs shape: {freqs.shape}, spec_arr shape: {spec_arr.shape}, trace_arr shape: {trace_arr.shape}, times_trace_arr shape: {times_trace_arr.shape}, snr_arr shape: {snr_arr.shape}, times shape: {times.shape}, run_no shape: {run_no.shape}  ")
-    print(f"trigger types: {np.unique(event_info['triggerType'])}")
-
-    return spec_arr, trace_arr, times_trace_arr, snr_arr, run_no, times, freqs, event_info, glitch_arr
-
-    
-#### Spectral analysis functions ####
-def find_amplitude_ratio_in_band(station_id, freqs, norm_spec_arr, upward_channels, downward_channels, reference_channels, freq_min, freq_max):
-    '''Find normalized amplitude ratio of upward vs downward channels in a given frequency band. This will be replaced by lab measurements in the future.'''
-    freq_mask = (freqs >= freq_min) & (freqs <= freq_max)
-    ratio_list = []
-    spectrum_list = []
-
-    ref_specs = np.stack([norm_spec_arr[ch][:, freq_mask] for ch in reference_channels], axis=0)
-    ref_spectra_per_ch = np.median(ref_specs, axis=2)   # (n_ref_ch, n_events)
-    ref_spec = np.median(ref_spectra_per_ch, axis=0) # (n_events,)
-
-    for ch in upward_channels + downward_channels:
-        masked_spec = norm_spec_arr[ch][:, freq_mask]        # (n_events, n_freqs)
-        masked_spec_med = np.median(masked_spec, axis=1)       # (n_events,)
-
-        spec_ratio = masked_spec_med / ref_spec
-        ratio_list.append(spec_ratio)
-        spectrum_list.append(masked_spec_med)
-
-    return ratio_list, spectrum_list
-
-
-def find_amplitude_ratio_in_band_specific_bkg(station_id, freqs, norm_spec_arr, upward_channels, downward_channels):
-    '''Find normalized amplitude ratio of upward vs downward channels in specific frequency bands. Backgrouns were defined using wiki page: https://radio.uchicago.edu/wiki/index.php/Features_observed_in_data'''
-
-    # Galactic excess frequency band
-    if station_id == 14:
-        reference_channels_gal = [12, 14, 19]
-    else:
-        reference_channels_gal = downward_channels
-    freq_min_gal = 80 * units.MHz
-    freq_max_gal = 120 * units.MHz
-    ratio_arr_gal, _ = find_amplitude_ratio_in_band(station_id, freqs, norm_spec_arr, upward_channels, downward_channels, reference_channels_gal, freq_min_gal, freq_max_gal)
-
-    # 360-380 MHz frequency band
-    reference_channels_360 = downward_channels
-    freq_min_360 = 360 * units.MHz
-    freq_max_360 = 380 * units.MHz
-    ratio_arr_360, _ = find_amplitude_ratio_in_band(station_id, freqs, norm_spec_arr, upward_channels, downward_channels, reference_channels_360, freq_min_360, freq_max_360)
-
-    # 482-485 MHz frequency band
-    reference_channels_482 = downward_channels
-    freq_min_482 = 482 * units.MHz
-    freq_max_485 = 485 * units.MHz
-    ratio_arr_482, _ = find_amplitude_ratio_in_band(station_id, freqs, norm_spec_arr, upward_channels, downward_channels, reference_channels_482, freq_min_482, freq_max_485)
-
-    reference_channels_240 = downward_channels
-    freq_min_240 = 240 * units.MHz
-    freq_max_272 = 272 * units.MHz
-    ratio_arr_240, _ = find_amplitude_ratio_in_band(station_id, freqs, norm_spec_arr, upward_channels, downward_channels, reference_channels_240, freq_min_240, freq_max_272)
-
-    return ratio_arr_gal, ratio_arr_360, ratio_arr_482, ratio_arr_240
-
-def excess_info_from_ratio(ratio_arr, alpha=0.01, freq_range_str=""):
-    '''Calculate excess information from amplitude ratios in frequency bands.'''
-
-    log_ratio = np.log10(np.asarray(ratio_arr))
-    median_log_ratio = np.median(log_ratio)
-    mean_log_ratio = np.mean(log_ratio)
-    frac_pos = np.mean(log_ratio > 0)/np.mean(log_ratio < 0) if np.mean(log_ratio < 0) != 0 else np.inf
-
-    k = np.sum(log_ratio > 0)
-    n = int(log_ratio.size)
-    result = binomtest(k, n, p=0.5, alternative="greater")
-    pval = result.pvalue
-    statistic = result.statistic
-    confidence_interval = result.proportion_ci(confidence_level=0.99)
-
-    print(f"Binomial test for frequency range {freq_range_str}: k={k}, n={n}, pval={pval:.3e}, statistic={statistic:.3f}, 99% CI={confidence_interval}")
-
-    if pval > alpha:
-        validation = "NO EXCESS"
-    else:
-        if confidence_interval.low > 0.75:
-            validation = "STRONG EXCESS"
-        elif confidence_interval.low > 0.55:
-            validation = "MODERATE EXCESS"
-        else: 
-            validation = "WEAK EXCESS"       
-
-    return {
-        "median_log_ratio": median_log_ratio,
-        "mean_log_ratio": mean_log_ratio,
-        "frac_pos": frac_pos,
-        "pval": pval,
-        "validation": validation
-    }
-
-def validate_excess_in_bands(ratio_arr_gal, ratio_arr_360, ratio_arr_482, ratio_arr_240, alpha=0.01):
-    '''Validate excess information from amplitude ratios in frequency bands.'''
-    gal_info = excess_info_from_ratio(ratio_arr_gal, alpha, "80–120 MHz")
-    freq360_info = excess_info_from_ratio(ratio_arr_360, alpha, "360–380 MHz")
-    freq482_info = excess_info_from_ratio(ratio_arr_482, alpha, "482–485 MHz")
-    freq240_info = excess_info_from_ratio(ratio_arr_240, alpha, "240–272 MHz")
-    return {
-        "galactic_excess": gal_info,
-        "freq_360_380MHz": freq360_info,
-        "freq_482_485MHz": freq482_info,
-        "freq_240_272MHz": freq240_info
-    }
 
 def plot_time_integrated_surface_spectra_unnormalized(station_id, spec_arr, freqs, upward_channels, downward_channels, save_location, run_label, trigger_label):
     '''Plot time-integrated surface channel spectra.'''
@@ -453,9 +115,10 @@ def plot_time_integrated_surface_spectra_unnormalized(station_id, spec_arr, freq
         plt.plot(freqs / units.MHz, spec_mean, label=f'Ch {ch} (down)', linestyle='--')
 
     plt.xlabel('Frequency [MHz]')
-    plt.xlim(0, 800)
+    plt.xlim(50, 800)
+    plt.ylim(0, 5)
     plt.ylabel('Amplitude Spectrum [V/GHz]')
-    plt.title('Time-Integrated Spectrum of Surface Channels  (FORCE Trigger)')
+    plt.title(f'Time-Integrated Spectrum of Surface Channels  ({trigger_label} Trigger)')
     plt.legend(loc="upper right", 
                frameon=True,
                fancybox=True,
@@ -463,7 +126,7 @@ def plot_time_integrated_surface_spectra_unnormalized(station_id, spec_arr, freq
                edgecolor="black")
     plt.grid()
     plt.tight_layout()
-    plt.savefig(os.path.join(save_location, f"{trigger_label}_time_integrated_surface_spectra_unnormalized_force_trigger_{station_id}_{run_label}.pdf"))
+    plt.savefig(os.path.join(save_location, f"{trigger_label}_time_integrated_surface_spectra_unnormalized_{station_id}_{run_label}.pdf"))
 
 
 def plot_time_integrated_surface_spectra_normalized(station_id, norm_spec_arr, freqs, upward_channels, downward_channels, save_location, run_label):
@@ -493,7 +156,8 @@ def plot_time_integrated_surface_spectra_normalized(station_id, norm_spec_arr, f
     plt.axvspan(500, 650, color=normcolor, alpha=0.3, label="_nolegend_")
 
     plt.xlabel('Frequency [MHz]')
-    plt.xlim(0, 800)
+    plt.xlim(50, 800)
+    plt.ylim(0, 5)
     plt.ylabel('Amplitude Spectrum [V/GHz]')
     plt.title('Time-Integrated Spectrum of Surface Channels (FORCE Trigger)')
     
@@ -536,9 +200,10 @@ def plot_time_integrated_deep_spectra(station_id, spec_arr, freqs, vpol_channels
         plt.plot(freqs / units.MHz, spec_mean, label=f'Ch {ch} (HPOL)', linestyle='--')
 
     plt.xlabel('Frequency [MHz]')
-    plt.xlim(0, 800)
+    plt.xlim(50, 800)
+    plt.ylim(0, 5)
     plt.ylabel('Amplitude Spectrum [V/GHz]')
-    plt.title('Time-Integrated Spectrum of Deep Channels')
+    plt.title(f'Time-Integrated Spectrum of Deep Channels ({trigger_label} Trigger)')
     plt.legend(loc="upper right", 
                frameon=True,
                fancybox=True,
@@ -546,7 +211,7 @@ def plot_time_integrated_deep_spectra(station_id, spec_arr, freqs, vpol_channels
                edgecolor="black")
     plt.grid()
     plt.tight_layout()
-    plt.savefig(os.path.join(save_location, f"{trigger_label}_time_integrated_deep_spectra_unnormalized_force_trigger_{station_id}_{run_label}.pdf"))
+    plt.savefig(os.path.join(save_location, f"{trigger_label}_time_integrated_deep_spectra_unnormalized_{station_id}_{run_label}.pdf"))
 
 #### SNR analysis functions ####
 def calculate_statistics_log_snr(snr_arr):
@@ -648,15 +313,15 @@ def print_outlier_summary(outlier_details):
         n_outliers = len(entries)
 
         if n_outliers == 0:
-            print(f"Channel {ch}: 0 outliers")
+            logger.info(f"Channel {ch}: 0 outliers")
             continue
 
         k_ch = entries[0]["k"]
-        print(f"Channel {ch}: {n_outliers} outliers (k = {k_ch:.2f})")
+        logger.info(f"Channel {ch}: {n_outliers} outliers (k = {k_ch:.2f})")
 
         for e in entries:
-            print(f"  - run {e['run']}, event {e['eventNumber']}, "
-                  f"|z| = {e['z_abs']:.2f} (delta = {e['z_minus_k']:.2f} above k)")
+            logger.info(f"  - run {e['run']}, event {e['eventNumber']}, "
+                         f"|z| = {e['z_abs']:.2f} (delta = {e['z_minus_k']:.2f} above k)")
             
             
 def choose_day_interval(times):
@@ -680,7 +345,7 @@ def choose_day_interval(times):
     else:
         return 60
 
-def plot_snr_against_time(station_id,times,snr_arr,flag,z_log,k_list,channels,nrows=12,ncols=2,day_interval=None):
+def plot_snr_against_time(station_id,times,snr_arr,flag,z_log,k_list,channels,save_location,run_label,nrows=12,ncols=2,day_interval=None):
     times = pd.to_datetime(times,utc=True)
     channels = list(channels)
     n_channels = len(channels)
@@ -850,7 +515,7 @@ def report_vrms_characteristics(modality_dict, tail_dict, channel_list):
             tail_frac = None
             if tail_frac is not None:
                 tail_label += f" (fraction: {tail_frac:.3f})"
-            print(f"Ch {ch:02d}: {modality} ({tail_label})")
+            logger.info(f"Ch {ch:02d}: {modality} ({tail_label})")
             continue
 
         if not np.isnan(skew_trim_h):
@@ -895,7 +560,7 @@ def report_vrms_characteristics(modality_dict, tail_dict, channel_list):
         # output summary
         if tail_frac is not None:
             tail_label += f" (fraction: {tail_frac:.3f})"
-        print(f"Ch {ch:02d}: {modality} ({tail_label})")
+        logger.info(f"Ch {ch:02d}: {modality} ({tail_label})")
 
 
 def plot_vrms_values_against_time(times, vrms_arr_all, channel_list, station_id, run_label, save_location, n_rows = 12, n_cols = 2, day_interval=5):
@@ -1341,7 +1006,6 @@ def create_result_csv_file(station_id, run_label, n_events_force, surface_channe
                 df_spec_val = "?"
             else:
                 if ch in downward_channels:
-                    print(f"spectral_validation: {spectral_validation}")
                     if spectral_validation == "NO EXCESS":
                         df_spec_val = "OK"
                     elif spectral_validation == "WEAK EXCESS":
@@ -1351,7 +1015,6 @@ def create_result_csv_file(station_id, run_label, n_events_force, surface_channe
                     else:
                         df_spec_val = "?"
                 elif ch in upward_channels:
-                    print(f"spectral_validation: {spectral_validation}")
                     if spectral_validation in ["STRONG EXCESS", "MODERATE EXCESS"]:
                         df_spec_val = "OK"
                     elif spectral_validation == "WEAK EXCESS":
@@ -1481,7 +1144,7 @@ def create_result_csv_file(station_id, run_label, n_events_force, surface_channe
     health_cols =["SNR", "Galaxy (FORCE)", "Vrms (FORCE)", "Vrms (LT)", "Vrms (RADIANT0)", "Vrms (RADIANT1)", "Glitching"]
     df["Channel Health"] = df[health_cols].apply(channel_health, axis=1)
     df.to_csv(out_csv_file, index=False)
-    print(f"Validation summary saved to {out_csv_file}")
+    logger.info(f"Validation summary saved to {out_csv_file}")
 
 
 if __name__ == "__main__":
@@ -1501,8 +1164,6 @@ if __name__ == "__main__":
     run_selection.add_argument("--time_range", nargs=2, type=str, metavar=("START_DATE", "END_DATE"),
                             help="Date range to analyze (inclusive). Provide start and end dates separated by a space in YYYY-MM-DD format, e.g. --time_range 2024-07-15 2024-09-30")
 
-
-    surface_channels, deep_channels, upward_channels, downward_channels, vpol_channels, hpol_channels, phased_array_channels, all_channels = station_channels(argparser.parse_args().station_id)
 
     args = argparser.parse_args()
 
@@ -1536,21 +1197,38 @@ if __name__ == "__main__":
     else:
         run_label = f"runs_{first_run}_{last_run}"
 
+    # Start logging
+    setup_logging(station_id, run_label)
+    logger.info(f"Starting analysis for station {station_id}, runs: {run_numbers}, backend: {backend}")
+
+    # Set the plotting style
+    set_plot_style()
 
     # Create save location directory if it doesn't exist
     save_location = os.path.expanduser(args.save_location)
     os.makedirs(save_location, exist_ok=True)
 
-    result_txt = os.path.join(save_location, f"result_station{station_id}_{run_label}.txt")
-    start_print_logging(result_txt)
+    # Get channel lists from config
+    config = get_station_config(station_id)
+    surface_channels = config["surface_channels"]
+    deep_channels = config["deep_channels"]
+    upward_channels = config["upward_channels"]
+    downward_channels = config["downward_channels"]
+    vpol_channels = config["vpol_channels"]
+    hpol_channels = config["hpol_channels"]
+    phased_array_channels = config["phased_array_channels"]
+    all_channels = config["all_channels"]
+    reference_channels_galaxy = config["reference_channels_galaxy"]
+    reference_channels = config["reference_channels"]
 
-    # Read data
+    # Read data 
+    ###### !!!!! Implement monitoring data !!!!! ######
     spec_arr, trace_arr, times_trace_arr, snr_arr, run_no, times, freqs, event_info, glitch_arr = read_rnog_data(station_id, run_numbers, backend=backend) 
 
     # Normalize surface channel spectra
     norm_spec_arr, scale_factors = normalize_channels(spec_arr, freqs, downward_channels, upward_channels)
-    print(f"Event info trigger type: {event_info['triggerType']}")
-    print(f"Spec arr shape: {spec_arr.shape}, Norm spec arr shape: {norm_spec_arr.shape}")
+    logger.debug(f"Event info trigger type: {event_info['triggerType']}")
+    logger.debug(f"Spec arr shape: {spec_arr.shape}, Norm spec arr shape: {norm_spec_arr.shape}")
 
     # Select FORCE trigger events
     force_mask = choose_trigger_type(event_info, "FORCE")
@@ -1558,37 +1236,58 @@ if __name__ == "__main__":
     # Spectral analysis for FORCE trigger events only
     spec_arr_force = spec_arr[:, force_mask, :]
     norm_spec_arr_force = norm_spec_arr[:, force_mask, :]
-    print(f"Number of FORCE trigger events: {spec_arr_force.shape[1]}")
+    logger.info(f"Number of FORCE trigger events: {spec_arr_force.shape[1]}")
 
     if len(spec_arr_force[1]) < 30:
-        print("!!!! Warning: Less than 30 FORCE-trigger events, results of the sign test may not be reliable. !!!!")
+        logger.warning("Less than 30 FORCE-trigger events, results of the sign test may not be reliable.")
 
-    ratio_arr_gal, ratio_arr_360, ratio_arr_482, ratio_arr_240 = find_amplitude_ratio_in_band_specific_bkg(station_id, freqs, norm_spec_arr_force, upward_channels, downward_channels)
+    band_config = copy.deepcopy(SPECTRAL_BANDS)
+
+    for band_name in band_config:
+        if band_name != "galactic_excess":
+            band_config[band_name]["reference_channels"] = reference_channels
+        elif band_name == "galactic_excess":
+            band_config[band_name]["reference_channels"] = reference_channels_galaxy
+        else:
+            logger.error(f"Unknown band name {band_name} in SPECTRAL_BANDS config")
+            raise ValueError(f"Unknown band name {band_name} in SPECTRAL_BANDS config")
+
+    logger.debug(f"Band configuration for spectral analysis: {band_config}")
+
+    ratio_arr_dict = find_amplitude_ratio_in_band_specific_bkg(freqs, norm_spec_arr_force, upward_channels, downward_channels, **band_config)
+
     channels_order = upward_channels + downward_channels
     ch_to_idx = {ch: i for i, ch in enumerate(channels_order)}
-    print(ch_to_idx)
+    logger.debug(f"Channel to index mapping: {ch_to_idx}")
 
+    all_excess_info = {}
     all_validation_results = {}
-    for ch in surface_channels:
-        print(f"\nAnalyzing channel {ch}")
-        i = ch_to_idx[ch]
-        print(i)
 
-        validation_results = validate_excess_in_bands(
-            ratio_arr_gal[i],
-            ratio_arr_360[i],
-            ratio_arr_482[i],
-            ratio_arr_240[i],
-            alpha=0.01
-        )
+    for ch in surface_channels:
+        logger.info(f"\nAnalyzing spectral excess for channel {ch}...")
+        i = ch_to_idx[ch]
+        ratio_arr_dict_ch = {}
+
+        for band_name, ratio_arr in ratio_arr_dict.items():
+            ratio_arr_dict_ch[band_name] = ratio_arr[i]
+
+        excess_info_results = excess_info_from_ratio_specific_bkg(ratio_arr_dict_ch, ALPHA_SPEC, CI_THRESHOLDS_SPEC)
+        validation_results = validate_excess_in_bands(excess_info_results)
+
+        all_excess_info[ch] = excess_info_results
         all_validation_results[ch] = validation_results
 
-        for band, results in validation_results.items():
-            print(f"=== {band} ===")
-            for key, value in results.items():
-                print(f"{key}: {value}")
-            print("==============")
+        for band, results in excess_info_results.items():
+            summary = "\n".join(
+                [f"{key}: {value}" for key, value in results.items()]
+            )
+            logger.info(
+                f"\n=== {band} ===\n"
+                f"{summary}\n"
+                f"=============="
+            )
 
+    # Select other trigger types and plot them
     lt_mask = choose_trigger_type(event_info, "LT")
     spec_arr_lt = spec_arr[:, lt_mask, :]
 
@@ -1620,12 +1319,12 @@ if __name__ == "__main__":
     times_force = times[force_mask]
 
     log_snr_arr, log_mean_list, log_median_list, log_std_list, log_difference_list = calculate_statistics_log_snr(snr_arr_force)
-    ref_log_mean_filename = f"station_{station_id}_ref_log_mean_snr.json"
-    ref_log_std_filename = f"station_{station_id}_ref_log_std_snr.json"
+    ref_log_mean_filename = f"expected_values/station_{station_id}_ref_log_mean_snr.json"
+    ref_log_std_filename = f"expected_values/station_{station_id}_ref_log_std_snr.json"
     ref_log_mean_list = load_values_json(ref_log_mean_filename)
     ref_log_std_list = load_values_json(ref_log_std_filename)
     z_score_arr_log_snr = calculate_z_score_snr(log_snr_arr, ref_log_mean_list, ref_log_std_list, all_channels)
-    k_values_filename_snr = f"station_{station_id}_k_ref_values_snr.json"
+    k_values_filename_snr = f"expected_values/station_{station_id}_k_ref_values_snr.json"
     k_values_log_snr = load_values_json(k_values_filename_snr)
     flag_outliers_snr = outlier_flag(z_score_arr_log_snr, k_values_log_snr, all_channels)
 
@@ -1633,38 +1332,38 @@ if __name__ == "__main__":
     print_outlier_summary(outlier_details_snr)
 
     day_interval = choose_day_interval(times)
-    plot_snr_against_time(station_id, times_force, snr_arr_force, flag_outliers_snr, z_score_arr_log_snr, k_values_log_snr, all_channels, nrows=12, ncols=2, day_interval=day_interval)
+    plot_snr_against_time(station_id, times_force, snr_arr_force, flag_outliers_snr, z_score_arr_log_snr, k_values_log_snr, all_channels, save_location, run_label, nrows=12, ncols=2, day_interval=day_interval)
 
     # Vrms analysis
     vrms_arr, vrms_arr_force, vrms_arr_radiant0, vrms_arr_radiant1, vrms_arr_lt = calculate_vrms(trace_arr, event_info)
-    print(f"Number of RADIANT0 trigger events: {len(vrms_arr_radiant0[1])}, Number of RADIANT1 trigger events: {len(vrms_arr_radiant1[1])}, Number of LT trigger events: {len(vrms_arr_lt[1])}")
+    logger.info(f"Number of RADIANT0 trigger events: {len(vrms_arr_radiant0[1])}, Number of RADIANT1 trigger events: {len(vrms_arr_radiant1[1])}, Number of LT trigger events: {len(vrms_arr_lt[1])}")
 
     modality_dict_force = kde_modality(vrms_arr_force, all_channels, bandwidth=None, grid_points=512, peak_prominence=0.01)
     tail_dict_force = tail_fraction_and_trimmed_skew_two_sided(vrms_arr_force, all_channels, lower_percentile=25, upper_percentile=75, extreme_k=3)
-    print("\nVrms characteristics for FORCE trigger events:")
+    logger.info("\nVrms characteristics for FORCE trigger events:")
     if len(vrms_arr_force[1]) < 100:
-            print(f"!!!! Warning: FORCE trigger has less than 100 valid Vrms entries ({len(vrms_arr_force)}). Results for the Vrms statistics may be unreliable. !!!!")
+            logger.warning(f"FORCE trigger has less than 100 valid Vrms entries ({len(vrms_arr_force)}). Results for the Vrms statistics may be unreliable.")
     report_vrms_characteristics(modality_dict_force, tail_dict_force, all_channels)
 
     modality_dict_radiant0 = kde_modality(vrms_arr_radiant0, all_channels, bandwidth=None, grid_points=512, peak_prominence=0.01)
     tail_dict_radiant0 = tail_fraction_and_trimmed_skew_two_sided(vrms_arr_radiant0, all_channels, lower_percentile=25, upper_percentile=75, extreme_k=3)
-    print("\nVrms characteristics for RADIANT0 trigger events:")
+    logger.info("\nVrms characteristics for RADIANT0 trigger events:")
     if len(vrms_arr_radiant0[1]) < 100:
-            print(f"!!!! Warning: RADIANT0 trigger has less than 100 valid Vrms entries ({len(vrms_arr_radiant0)}). Results for the Vrms statistics may be unreliable. !!!!")
+            logger.warning(f"RADIANT0 trigger has less than 100 valid Vrms entries ({len(vrms_arr_radiant0)}). Results for the Vrms statistics may be unreliable.")
     report_vrms_characteristics(modality_dict_radiant0, tail_dict_radiant0, all_channels)
 
     modality_dict_radiant1 = kde_modality(vrms_arr_radiant1, all_channels, bandwidth=None, grid_points=512, peak_prominence=0.01)
     tail_dict_radiant1 = tail_fraction_and_trimmed_skew_two_sided(vrms_arr_radiant1, all_channels, lower_percentile=25, upper_percentile=75, extreme_k=3)
-    print("\nVrms characteristics for RADIANT1 trigger events:")
+    logger.info("\nVrms characteristics for RADIANT1 trigger events:")
     if len(vrms_arr_radiant1[1]) < 100:
-            print(f"!!!! Warning: RADIANT1 trigger has less than 100 valid Vrms entries ({len(vrms_arr_radiant1)}). Results for the Vrms statistics may be unreliable. !!!!")
+            logger.warning(f"RADIANT1 trigger has less than 100 valid Vrms entries ({len(vrms_arr_radiant1)}). Results for the Vrms statistics may be unreliable.")
     report_vrms_characteristics(modality_dict_radiant1, tail_dict_radiant1, all_channels)
 
     modality_dict_lt = kde_modality(vrms_arr_lt, all_channels, bandwidth=None, grid_points=512, peak_prominence=0.01)
     tail_dict_lt = tail_fraction_and_trimmed_skew_two_sided(vrms_arr_lt, all_channels, lower_percentile=25, upper_percentile=75, extreme_k=3)
-    print("\nVrms characteristics for LT trigger events:")  
+    logger.info("\nVrms characteristics for LT trigger events:")  
     if len(vrms_arr_lt[1]) < 100:
-            print(f"!!!! Warning: LT trigger has less than 100 valid Vrms entries ({len(vrms_arr_lt)}). Results for the Vrms statistics may be unreliable. !!!!")
+            logger.warning(f"LT trigger has less than 100 valid Vrms entries ({len(vrms_arr_lt)}). Results for the Vrms statistics may be unreliable.")
     report_vrms_characteristics(modality_dict_lt, tail_dict_lt, all_channels)
     plot_vrms_values_against_time(times, vrms_arr, all_channels, station_id, run_label, save_location, n_rows=12, n_cols=2, day_interval=day_interval)
 
@@ -1676,18 +1375,22 @@ if __name__ == "__main__":
     
     # Glitching analysis
     glitch_info = binomtest_glitch_fraction(glitch_arr, all_channels, alpha=0.01)
-    print("\nGlitching analysis results:")
-    for ch in all_channels:
-        info = glitch_info[ch]
-        print(
+    lines = [
+        (
             f"Channel {ch:2d} | "
-            f"n_glitches: {info['n_glitches']:4d} | "
-            f"n_events: {info['n_events']:<4d} | "
-            f"(frac={info['glitch_fraction']:.3f}) | "
-            f"p={info['pval']:.2e} | "
-            f"CI99%={info['confidence_interval']} | "
-            f"{info['validation']}"
+            f"n_glitches: {glitch_info[ch]['n_glitches']:4d} | "
+            f"n_events: {glitch_info[ch]['n_events']:<4d} | "
+            f"(frac={glitch_info[ch]['glitch_fraction']:.3f}) | "
+            f"p={glitch_info[ch]['pval']:.2e} | "
+            f"CI99%={glitch_info[ch]['confidence_interval']} | "
+            f"{glitch_info[ch]['validation']}"
         )
+        for ch in all_channels]
+    
+    logger.info(
+        "\nGlitching analysis results:\n" +
+        "\n".join(lines)
+    )
     glitching_violin_plot(glitch_arr, all_channels, station_id, run_label, save_location)
     plot_glitch_q99_over_time(np.array(times), glitch_arr, all_channels, station_id, run_label, save_location)
 
@@ -1703,7 +1406,6 @@ if __name__ == "__main__":
         debug_plot_snr_distribution(log_snr_arr, channel_list=all_channels, save_location=save_location, station_id=station_id, run_label=run_label, bins=30)
         debug_plot_z_score_snr(z_score_arr_log_snr, channel_list=all_channels, save_location=save_location, station_id=station_id, run_label=run_label, bins=30)   
         
-    stop_print_logging()
 
     # Create summary CSV file
     n_events_force = spec_arr_force.shape[1]
